requirements.txt
numpy~=1.20.2
flask~=1.1.2
scikit-learn~=0.24.2
pandas~=1.2.4
pillow~=8.2.0
matplotlib~=3.4.2